moved disclaimer modal-related code from mainLayouts.tsx to disclaimer.tsx.
